chnged uint32_t in all functions

+++ DebugWorkshop/ImagesAndWords/Images.cpp
@@ -29,7 +29,7 @@ bool WriteImage(Image* img, const char* filename)
 	os.write((char*)&img->header, sizeof(ImageHeader));
 
 	// calculate image size
-	uint16_t imgsize = img->header.width * img->header.height;
+	uint32_t imgsize = img->header.width * img->header.height;
 
 	// write image content
 	os.write(img->data, imgsize);
@@ -78,7 +78,7 @@ Image *GenerateDummyImage(uint16_t width, uint16_t height)
 	img->header.mode = 1;
 
 	// calculate image size
-	uint16_t imgsize = img->header.width * img->header.height;
+	uint32_t imgsize = img->header.width * img->header.height;
 	img->data = new char[imgsize];
 
 	// generate dummy content
